Add play_game method  to play a round
There is some experimental stuff commented out. The play_gamecode seems to work. Now it's just a case of making it more general by moving the code to methods and call these as needed.

Have to keep in mind that if I want to use player variables I need to call these to use ion other class methods. Might consider figuring out if this works by adding the players as variables for initializing the game class.